secured server Done
